refactor: move ranking api from meetups to users
/meetups/ranking -> /users/ranking

#15
